Initial support for `pip-repositories` configuration
