+++ conda_lock/models/pip_repository.py
@@ -1,0 +1,5 @@
+from conda_lock.models.package_source import PackageSource
+
+
+class PipRepository(PackageSource):
+    pass
